+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Fix Player: stabilize AudioSessionRouteChangeMonitor tests and add macOS guard to extension.
+
+### Changed
+- Auto-resume playback after short conenction drops over Bluetooth. (Player)
 
 ## [0.7.2] - 2025-09-24